fix structure appel api
