+++ src/api/BasePublic.restAPI.ts
@@ -1,0 +1,7 @@
+import { BaseRestApi } from './Base.restAPi';
+
+export abstract class BasePublicRestApi extends BaseRestApi {
+	protected header = new Headers({
+		'Content-Type': 'application/json'
+	});
+}
